refact: enums 패키지를 global 하위로 이동
