+++ src/main/java/umc/cockple/demo/domain/member/domain/Member.java
@@ -7,9 +7,9 @@ import umc.cockple.demo.domain.bookmark.domain.ExerciseBookmark;
 import umc.cockple.demo.domain.bookmark.domain.PartyBookmark;
 import umc.cockple.demo.domain.contest.domain.Contest;
 import umc.cockple.demo.domain.notification.domain.Notification;
-import umc.cockple.demo.enums.Gender;
-import umc.cockple.demo.enums.Level;
-import umc.cockple.demo.enums.MemberStatus;
+import umc.cockple.demo.global.enums.Gender;
+import umc.cockple.demo.global.enums.Level;
+import umc.cockple.demo.global.enums.MemberStatus;
 import umc.cockple.demo.mapping.MemberExercise;
 import umc.cockple.demo.mapping.MemberParty;
 import umc.cockple.demo.global.common.BaseEntity;
