--- src/main/java/umc/cockple/demo/enums/ActiveDay.java
@@ -1,12 +1,0 @@
-package umc.cockple.demo.enums;
-
-public enum ActiveDay {
-
-    SUNDAY,
-    MONDAY,
-    TUESDAY,
-    WEDNESDAY,
-    THURSDAY,
-    FRIDAY,
-    SATURDAY
-}
