--- src/main/java/umc/cockple/demo/enums/ActivityTime.java
@@ -1,9 +1,0 @@
-package umc.cockple.demo.enums;
-
-public enum ActivityTime {
-
-    MORNING,
-    AFTERNOON,
-    ALWAYS
-
-}
