--- src/main/java/umc/cockple/demo/enums/ExerciseOrderType.java
@@ -1,6 +1,0 @@
-package umc.cockple.demo.enums;
-
-public enum ExerciseOrderType {
-    LATEST,
-    EARLIEST
-}
