--- src/main/java/umc/cockple/demo/enums/Gender.java
@@ -1,5 +1,0 @@
-package umc.cockple.demo.enums;
-
-public enum Gender {
-    MALE, FEMALE
-}
