--- src/main/java/umc/cockple/demo/enums/Level.java
@@ -1,15 +1,0 @@
-package umc.cockple.demo.enums;
-
-public enum Level {
-
-    EXPERT, // 자강
-    SEMI_EXPERT, // 준자강
-    A,
-    B,
-    C,
-    D,
-    BEGINNER, // 초심
-    NOVICE, // 왕초심
-    NONE // 급수없음
-
-}
