--- src/main/java/umc/cockple/demo/enums/MedalType.java
@@ -1,8 +1,0 @@
-package umc.cockple.demo.enums;
-
-public enum MedalType {
-
-    GOLD,
-    SILVER,
-    BRONZE
-}
