--- src/main/java/umc/cockple/demo/enums/MemberPartyStatus.java
@@ -1,8 +1,0 @@
-package umc.cockple.demo.enums;
-
-public enum MemberPartyStatus {
-
-    WAITING, // 수락대기
-
-    ACCEPT // 승인완료
-}
