--- src/main/java/umc/cockple/demo/enums/MemberStatus.java
@@ -1,6 +1,0 @@
-package umc.cockple.demo.enums;
-
-public enum MemberStatus {
-    // 회원 탈퇴시 INACTIVE
-    ACTIVE, INACTIVE
-}
