--- src/main/java/umc/cockple/demo/enums/NotificationType.java
@@ -1,5 +1,0 @@
-package umc.cockple.demo.enums;
-
-public enum NotificationType {
-    // 미정
-}
