--- src/main/java/umc/cockple/demo/enums/ParticipationType.java
@@ -1,12 +1,0 @@
-package umc.cockple.demo.enums;
-
-public enum ParticipationType {
-
-    SINGLE, // 단식
-
-    WOMEN_DOUBLES, // 여복
-
-    MEN_DOUBLES, // 남복
-
-    MIX_DOUBLES // 혼복
-}
