--- src/main/java/umc/cockple/demo/enums/PartyOrderType.java
@@ -1,9 +1,0 @@
-package umc.cockple.demo.enums;
-
-public enum PartyOrderType {
-
-    LATEST,
-    EARLIEST,
-    MANY_EXERCISE
-
-}
