--- src/main/java/umc/cockple/demo/enums/Role.java
@@ -1,8 +1,0 @@
-package umc.cockple.demo.enums;
-
-public enum Role {
-
-    party_MEMBER,
-
-    party_MANAGER
-}
