+++ src/main/java/umc/cockple/demo/global/enums/ActiveDay.java
@@ -1,0 +1,12 @@
+package umc.cockple.demo.global.enums;
+
+public enum ActiveDay {
+
+    SUNDAY,
+    MONDAY,
+    TUESDAY,
+    WEDNESDAY,
+    THURSDAY,
+    FRIDAY,
+    SATURDAY
+}
