+++ src/main/java/umc/cockple/demo/global/enums/ActivityTime.java
@@ -1,0 +1,9 @@
+package umc.cockple.demo.global.enums;
+
+public enum ActivityTime {
+
+    MORNING,
+    AFTERNOON,
+    ALWAYS
+
+}
