+++ src/main/java/umc/cockple/demo/global/enums/ExerciseOrderType.java
@@ -1,0 +1,6 @@
+package umc.cockple.demo.global.enums;
+
+public enum ExerciseOrderType {
+    LATEST,
+    EARLIEST
+}
