+++ src/main/java/umc/cockple/demo/global/enums/Gender.java
@@ -1,0 +1,5 @@
+package umc.cockple.demo.global.enums;
+
+public enum Gender {
+    MALE, FEMALE
+}
