+++ src/main/java/umc/cockple/demo/global/enums/Keyword.java
@@ -1,0 +1,10 @@
+package umc.cockple.demo.global.enums;
+
+public enum Keyword {
+
+    BRAND, // 브랜드 스폰
+    FREE, // 가입비 무료
+    FRIENDSHIP, // 친목
+    MANAGER_MATCH, // 운영진이 게임을 짜드려요
+    NONE // 선택안함
+}
