+++ src/main/java/umc/cockple/demo/global/enums/MedalType.java
@@ -1,0 +1,8 @@
+package umc.cockple.demo.global.enums;
+
+public enum MedalType {
+
+    GOLD,
+    SILVER,
+    BRONZE
+}
