+++ src/main/java/umc/cockple/demo/global/enums/MemberPartyStatus.java
@@ -1,0 +1,8 @@
+package umc.cockple.demo.global.enums;
+
+public enum MemberPartyStatus {
+
+    WAITING, // 수락대기
+
+    ACCEPT // 승인완료
+}
