+++ src/main/java/umc/cockple/demo/global/enums/MemberStatus.java
@@ -1,0 +1,6 @@
+package umc.cockple.demo.global.enums;
+
+public enum MemberStatus {
+    // 회원 탈퇴시 INACTIVE
+    ACTIVE, INACTIVE
+}
