+++ src/main/java/umc/cockple/demo/global/enums/NotificationType.java
@@ -1,0 +1,5 @@
+package umc.cockple.demo.global.enums;
+
+public enum NotificationType {
+    // 미정
+}
