+++ src/main/java/umc/cockple/demo/global/enums/ParticipationType.java
@@ -1,0 +1,12 @@
+package umc.cockple.demo.global.enums;
+
+public enum ParticipationType {
+
+    SINGLE, // 단식
+
+    WOMEN_DOUBLES, // 여복
+
+    MEN_DOUBLES, // 남복
+
+    MIX_DOUBLES // 혼복
+}
