+++ src/main/java/umc/cockple/demo/global/enums/PartyOrderType.java
@@ -1,0 +1,9 @@
+package umc.cockple.demo.global.enums;
+
+public enum PartyOrderType {
+
+    LATEST,
+    EARLIEST,
+    MANY_EXERCISE
+
+}
