+++ src/main/java/umc/cockple/demo/global/enums/Role.java
@@ -1,0 +1,8 @@
+package umc.cockple.demo.global.enums;
+
+public enum Role {
+
+    party_MEMBER,
+
+    party_MANAGER
+}
