+++ src/main/java/umc/cockple/demo/mapping/MemberExercise.java
@@ -4,7 +4,7 @@ import jakarta.persistence.*;
 import lombok.*;
 import umc.cockple.demo.domain.exercise.domain.Exercise;
 import umc.cockple.demo.domain.member.domain.Member;
-import umc.cockple.demo.enums.ExerciseOrderType;
+import umc.cockple.demo.global.enums.ExerciseOrderType;
 import umc.cockple.demo.global.common.BaseEntity;
 
 import java.time.LocalDateTime;
