+++ src/main/java/umc/cockple/demo/mapping/MemberParty.java
@@ -4,9 +4,9 @@ import jakarta.persistence.*;
 import lombok.*;
 import umc.cockple.demo.domain.party.domain.Party;
 import umc.cockple.demo.domain.member.domain.Member;
-import umc.cockple.demo.enums.PartyOrderType;
-import umc.cockple.demo.enums.MemberPartyStatus;
-import umc.cockple.demo.enums.Role;
+import umc.cockple.demo.global.enums.PartyOrderType;
+import umc.cockple.demo.global.enums.MemberPartyStatus;
+import umc.cockple.demo.global.enums.Role;
 import umc.cockple.demo.global.common.BaseEntity;
 
 import java.time.LocalDateTime;
